spanconfigstore: pull out the span config storage into internal struct
This commit pulls out what used to be the spanconfig.Store logic into
as separate spanconfigstore.spanConfigStore struct to exclusively track
span <-> config pairs. This also makes room for adding another internal
struct to track system span configs in a future patch.

Release note: None
